+++ plugin/ajax/import.php
@@ -1,0 +1,6 @@
+<?php
+sleep(2);
+
+// echo filter_input(INPUT_GET, 'itemtype');
+
+echo "OK";